Introduce AbiDecoder::GetAllDataRead.
When an AbiDecoder is constructed based on another one (e.g. using
ReadDynamic), it will be passed the entire tail of data, and it won't
be known where exactly the end of its data is.

Sometimes it is useful to get the exact piece of data corresponding to
some encoded fields.  The new method AbiDecoder::GetAllDataRead can
do that now, after all the relevant data has been read from the decoder.

To implement this, we keep track of what data has been read both
in the head and tail part, so that it works also recursively if we
decode other dynamic data structures from the data.